feature-initStudentAPP 1.新增student app 2.更改程式到對應的app
