+++ SecurityVulnerablilities/SecurityVulnerablilities/settings.py
@@ -41,6 +41,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'accounts',
+    'students',
 ]
 
 MIDDLEWARE = [
