+++ SecurityVulnerablilities/students/admin.py
@@ -1,0 +1,12 @@
+from django.contrib import admin
+from .models import Grade, Board
+
+class GradeAdmin(admin.ModelAdmin):
+	list_display = ("id", "student", "semester", "subject", "grade")
+	list_filter = ("semester", "subject")
+	
+class  BoardAdmin(admin.ModelAdmin):
+    list_display = ("id", "author", "content")
+
+admin.site.register(Grade, GradeAdmin)
+admin.site.register(Board, BoardAdmin)
