+++ SecurityVulnerablilities/templates/accounts/grade_search.html
@@ -44,7 +44,7 @@
             <tr>
                 <td>{{ result.semester }}</td>
                 <td>{{ result.studentId }} - {{ result.username }}</td>
-                <td>{{ result.get_subject_display }}</td>
+                <td>{{ result.subject }}</td>
                 <td>{{ result.grade }}</td>
             </tr>
         {% endfor %}
